Extract buildRequestBody into its own file
Pure mechanical move of the 60-line buildRequestBody function out of
useImagesInfinite.ts into buildRequestBody.ts. The function is re-exported
from useImagesInfinite so existing imports continue to work. No logic
changes.
